refactor(mcp): split http test utils
